gsi/ssdperf/requirements.txt
h5py
weaviate-client

